# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rochelle Ataide Confeitaria ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock:
    get:
      operationId: listStockItems
      tags: [stock]
      summary: List all stock items / ingredients
      parameters:
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Stock items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockItem"
    post:
      operationId: createStockItem
      tags: [stock]
      summary: Create a stock item / ingredient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
  /stock/{id}:
    get:
      operationId: getStockItem
      tags: [stock]
      summary: Get stock item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
    patch:
      operationId: updateStockItem
      tags: [stock]
      summary: Update stock item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
    delete:
      operationId: deleteStockItem
      tags: [stock]
      summary: Delete stock item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /stock/{id}/movement:
    post:
      operationId: addStockMovement
      tags: [stock]
      summary: Add stock movement (entry/exit)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockMovementInput"
      responses:
        "201":
          description: Movement recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"

  # ─── RECIPES ──────────────────────────────────────────────────────────────
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List all recipes / fichas técnicas
      responses:
        "200":
          description: Recipes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get recipe by ID with cost calculation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe with cost
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
    patch:
      operationId: updateRecipe
      tags: [recipes]
      summary: Update recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FINANCIAL ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    StockItem:
      type: object
      required: [id, name, unit, quantity, minStock, unitCost, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        unit:
          type: string
        quantity:
          type: number
        minStock:
          type: number
        unitCost:
          type: number
        supplier:
          type: ["string", "null"]
        isLow:
          type: boolean
        createdAt:
          type: string
    StockItemInput:
      type: object
      required: [name, unit, quantity, minStock, unitCost]
      properties:
        name:
          type: string
        unit:
          type: string
        quantity:
          type: number
        minStock:
          type: number
        unitCost:
          type: number
        supplier:
          type: string
    StockItemUpdate:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
        quantity:
          type: number
        minStock:
          type: number
        unitCost:
          type: number
        supplier:
          type: string
    StockMovementInput:
      type: object
      required: [type, quantity]
      properties:
        type:
          type: string
          enum: [entry, exit]
        quantity:
          type: number
        reason:
          type: string

    # ─── RECIPE ─────────────────────────────────────────────────────────────
    RecipeIngredient:
      type: object
      required: [stockItemId, stockItemName, quantity, unit, cost]
      properties:
        stockItemId:
          type: integer
        stockItemName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        cost:
          type: number
    RecipeIngredientInput:
      type: object
      required: [stockItemId, quantity, unit]
      properties:
        stockItemId:
          type: integer
        quantity:
          type: number
        unit:
          type: string
    Recipe:
      type: object
      required: [id, productId, productName, yield, prepTime, totalCost, unitCost, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        yield:
          type: integer
        prepTime:
          type: integer
        instructions:
          type: ["string", "null"]
        totalCost:
          type: number
        unitCost:
          type: number
        cmvPercent:
          type: ["number", "null"]
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        createdAt:
          type: string
    RecipeInput:
      type: object
      required: [productId, yield, prepTime, ingredients]
      properties:
        productId:
          type: integer
        yield:
          type: integer
        prepTime:
          type: integer
        instructions:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredientInput"
    RecipeUpdate:
      type: object
      properties:
        yield:
          type: integer
        prepTime:
          type: integer
        instructions:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredientInput"

    # ─── FINANCIAL ──────────────────────────────────────────────────────────
...